+++ renderer/src/scene.cpp
@@ -375,7 +375,7 @@ template <template <typename> class VectorType>
 
 bool Scene<VectorType>::makeSurfaceDirectConnection(const VectorType<Float> &p1, const VectorType<Float> &p2, const Float &scaling, smp::Sampler &sampler,
 														Float &distTravelled, VectorType<Float> &dirToSensor, Float &distToSensor, Float &weight,
-														scn::NEECostFunction<VectorType> &costFunction, Problem &problem, Float *initialization) const{
+														scn::NEECostFunction<VectorType> &costFunction, Problem &problem, double *initialization) const{
 
 	Matrix3x3 dpdv0((Float)0);
 	Matrix3x3 dvdv0((Float)1, 0, 0,
@@ -937,7 +937,7 @@ template <template <typename> class VectorType>
 void Scene<VectorType>::addEnergy(image::SmallImage &img,
 			const VectorType<Float> &p, const VectorType<Float> &d, Float distTravelled, int &depth,
 			Float val, const med::Medium &medium, smp::Sampler &sampler, const Float& scaling,
-			scn::NEECostFunction<VectorType> &costFunction, Problem &problem, Float *initialization) const {
+			scn::NEECostFunction<VectorType> &costFunction, Problem &problem, double *initialization) const {
 
 #ifdef USE_WEIGHT_NORMALIZATION
 	val *=	static_cast<Float>(img.getXRes()) * static_cast<Float>(img.getYRes())
